college_management/requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
dj-database-url==3.0.1
Django==5.2.4
django-admin-extra-buttons==2.1.1
django-autocomplete-light==3.12.1
django-import-export==4.3.8
django-simple-history==3.10.1
django-stubs==5.2.1
django-stubs-ext==5.2.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
tablib==3.8.0
types-PyYAML==6.0.12.20250516
typing_extensions==4.14.0
whitenoise==6.9.0
